Visual Studio project for SkelCL. All tests compiling and passing.

+++ test/MapTests.cpp
@@ -131,7 +131,7 @@ TEST_F(MapTest, SimpleMultiDeviceMap) {
 TEST_F(MapTest, SimpleMultiDeviceMap2) {
   skelcl::terminate();
   skelcl::init(skelcl::nDevices(2));
-  skelcl::Map<int(float)> m{ "float func(float f) \
+  skelcl::Map<int(float)> m{ "int func(float f) \
     { return skelcl_get_device_id(); }" };
 
   skelcl::Vector<float> input(10);
